requirements.txt
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.118.0
matplotlib>=3.8.0
numpy>=1.26.0
Pillow>=10.0.0
requests>=2.31.0
